task3/requirements.txt
javaobj-py3==0.2.4
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycryptodome==3.5.1
pyjks==17.1.1
twofish==0.3.0
